- [x] Link to issue, e.g. `Resolves #NNN`
 - [x] Documentation added (if applicable)
 - [ ] Tests added
 - [x] Branch rebased on top of current main (`git pull --rebase origin main`)
 - [ ] Changes squashed to a single commit (described [here](http://gitready.com/advanced/2009/02/10/squashing-commits-with-rebase.html))
 - [ ] Build is green in Travis CI
 - [x] You have certified that the contribution is your original work and that you license the work to the project under the [Apache 2 license](https://github.com/nats-io/nats-server/blob/main/LICENSE)

Resolves #4341

### Changes proposed in this pull request:

The sublist cache currently uses a map[string]*SublistResult to store the values. 
A go map is obviously not limited in size so the current implementation runs reduceCacheCount() to remove random elements from the cache once there are slCacheMax = 1024 elemnts in it. 
It remove elements untill there are only slCacheSweep = 256 left in it. This becomes really expensive when there are many subscriptions and the hit rate is low because the cache just permanently deletes elements from the cache and locks for a relatively long time.

I experimented with a simple fixed size cache implementation that is just backed by a fixed size array.
Its also lock free using atomic.Pointer. I haven't implemented a collision strategy jet (not sure its needed for this cache).

```go
type Cache struct {
	data     [slCacheMax]atomic.Pointer[cacheEntry]
	hashFunc func(string) int
}

type cacheEntry struct {
	key   string
	value *SublistResult
}
```

This implementation ensures that there are never more than slCacheMax elements in the cache with zero overhead.
Its also faster in all of my tests. I still need to add unit tests but alread wanted to drop it here for discussion. 

Nats bench results in my Ryzen 9 5900HX:

<details>
 <summary>Old Cache</summary>

Old cache slCacheMax 1024:
nats bench foo --msgs 1000000 --pub 100 --no-progress --multisubject --multisubjectmax 100000
 Pub stats: 1,465,732 msgs/sec ~ 178.92 MB/sec
 [1] 27,027 msgs/sec ~ 3.30 MB/sec (10000 msgs)
 [2] 25,414 msgs/sec ~ 3.10 MB/sec (10000 msgs)
 [3] 25,258 msgs/sec ~ 3.08 MB/sec (10000 msgs)
 [4] 24,885 msgs/sec ~ 3.04 MB/sec (10000 msgs)
 [5] 24,929 msgs/sec ~ 3.04 MB/sec (10000 msgs)
 [6] 23,350 msgs/sec ~ 2.85 MB/sec (10000 msgs)
 [7] 22,923 msgs/sec ~ 2.80 MB/sec (10000 msgs)
 [8] 22,887 msgs/sec ~ 2.79 MB/sec (10000 msgs)
 [9] 22,801 msgs/sec ~ 2.78 MB/sec (10000 msgs)
 [10] 21,709 msgs/sec ~ 2.65 MB/sec (10000 msgs)
 [11] 21,549 msgs/sec ~ 2.63 MB/sec (10000 msgs)
 [12] 21,456 msgs/sec ~ 2.62 MB/sec (10000 msgs)
 [13] 21,184 msgs/sec ~ 2.59 MB/sec (10000 msgs)
 [14] 20,086 msgs/sec ~ 2.45 MB/sec (10000 msgs)
 [15] 19,926 msgs/sec ~ 2.43 MB/sec (10000 msgs)
 [16] 19,687 msgs/sec ~ 2.40 MB/sec (10000 msgs)
 [17] 19,500 msgs/sec ~ 2.38 MB/sec (10000 msgs)
 [18] 19,379 msgs/sec ~ 2.37 MB/sec (10000 msgs)
 [19] 19,370 msgs/sec ~ 2.36 MB/sec (10000 msgs)
 [20] 19,321 msgs/sec ~ 2.36 MB/sec (10000 msgs)
 [21] 19,313 msgs/sec ~ 2.36 MB/sec (10000 msgs)
 [22] 19,315 msgs/sec ~ 2.36 MB/sec (10000 msgs)
 [23] 19,257 msgs/sec ~ 2.35 MB/sec (10000 msgs)
 [24] 19,192 msgs/sec ~ 2.34 MB/sec (10000 msgs)
 [25] 19,190 msgs/sec ~ 2.34 MB/sec (10000 msgs)
 [26] 19,175 msgs/sec ~ 2.34 MB/sec (10000 msgs)
 [27] 19,082 msgs/sec ~ 2.33 MB/sec (10000 msgs)
 [28] 18,537 msgs/sec ~ 2.26 MB/sec (10000 msgs)
 [29] 18,517 msgs/sec ~ 2.26 MB/sec (10000 msgs)
 [30] 18,484 msgs/sec ~ 2.26 MB/sec (10000 msgs)
 [31] 18,395 msgs/sec ~ 2.25 MB/sec (10000 msgs)
 [32] 18,305 msgs/sec ~ 2.23 MB/sec (10000 msgs)
 [33] 18,196 msgs/sec ~ 2.22 MB/sec (10000 msgs)
 [34] 18,029 msgs/sec ~ 2.20 MB/sec (10000 msgs)
 [35] 18,028 msgs/sec ~ 2.20 MB/sec (10000 msgs)
 [36] 18,027 msgs/sec ~ 2.20 MB/sec (10000 msgs)
 [37] 17,820 msgs/sec ~ 2.18 MB/sec (10000 msgs)
 [38] 17,660 msgs/sec ~ 2.16 MB/sec (10000 msgs)
 [39] 17,599 msgs/sec ~ 2.15 MB/sec (10000 msgs)
 [40] 17,387 msgs/sec ~ 2.12 MB/sec (10000 msgs)
 [41] 17,165 msgs/sec ~ 2.10 MB/sec (10000 msgs)
 [42] 17,104 msgs/sec ~ 2.09 MB/sec (10000 msgs)
 [43] 16,919 msgs/sec ~ 2.07 MB/sec (10000 msgs)
 [44] 16,943 msgs/sec ~ 2.07 MB/sec (10000 msgs)
 [45] 16,916 msgs/sec ~ 2.06 MB/sec (10000 msgs)
 [46] 16,942 msgs/sec ~ 2.07 MB/sec (10000 msgs)
 [47] 16,934 msgs/sec ~ 2.07 MB/sec (10000 msgs)
 [48] 16,890 msgs/sec ~ 2.06 MB/sec (10000 msgs)
 [49] 16,856 msgs/sec ~ 2.06 MB/sec (10000 msgs)
 [50] 16,855 msgs/sec ~ 2.06 MB/sec (10000 msgs)
 [51] 16,831 msgs/sec ~ 2.05 MB/sec (10000 msgs)
 [52] 16,816 msgs/sec ~ 2.05 MB/sec (10000 msgs)
 [53] 16,811 msgs/sec ~ 2.05 MB/sec (10000 msgs)
 [54] 16,484 msgs/sec ~ 2.01 MB/sec (10000 msgs)
 [55] 16,353 msgs/sec ~ 2.00 MB/sec (10000 msgs)
 [56] 16,307 msgs/sec ~ 1.99 MB/sec (10000 msgs)
 [57] 16,294 msgs/sec ~ 1.99 MB/sec (10000 msgs)
 [58] 16,248 msgs/sec ~ 1.98 MB/sec (10000 msgs)
 [59] 16,225 msgs/sec ~ 1.98 MB/sec (10000 msgs)
 [60] 16,134 msgs/sec ~ 1.97 MB/sec (10000 msgs)
 [61] 16,038 msgs/sec ~ 1.96 MB/sec (10000 msgs)
 [62] 15,964 msgs/sec ~ 1.95 MB/sec (10000 msgs)
 [63] 15,757 msgs/sec ~ 1.92 MB/sec (10000 msgs)
 [64] 15,728 msgs/sec ~ 1.92 MB/sec (10000 msgs)
 [65] 15,691 msgs/sec ~ 1.92 MB/sec (10000 msgs)
 [66] 15,676 msgs/sec ~ 1.91 MB/sec (10000 msgs)
 [67] 15,663 msgs/sec ~ 1.91 MB/sec (10000 msgs)
 [68] 15,640 msgs/sec ~ 1.91 MB/sec (10000 msgs)
 [69] 15,651 msgs/sec ~ 1.91 MB/sec (10000 msgs)
 [70] 15,606 msgs/sec ~ 1.91 MB/sec (10000 msgs)
 [71] 15,586 msgs/sec ~ 1.90 MB/sec (10000 msgs)
 [72] 15,569 msgs/sec ~ 1.90 MB/sec (10000 msgs)
 [73] 15,566 msgs/sec ~ 1.90 MB/sec (10000 msgs)
 [74] 15,538 msgs/sec ~ 1.90 MB/sec (10000 msgs)
 [75] 15,472 msgs/sec ~ 1.89 MB/sec (10000 msgs)
 [76] 15,466 msgs/sec ~ 1.89 MB/sec (10000 msgs)
 [77] 15,339 msgs/sec ~ 1.87 MB/sec (10000 msgs)
 [78] 15,245 msgs/sec ~ 1.86 MB/sec (10000 msgs)
 [79] 15,238 msgs/sec ~ 1.86 MB/sec (10000 msgs)
 [80] 15,227 msgs/sec ~ 1.86 MB/sec (10000 msgs)
 [81] 15,143 msgs/sec ~ 1.85 MB/sec (10000 msgs)
 [82] 15,127 msgs/sec ~ 1.85 MB/sec (10000 msgs)
 [83] 15,091 msgs/sec ~ 1.84 MB/sec (10000 msgs)
 [84] 15,035 msgs/sec ~ 1.84 MB/sec (10000 msgs)
 [85] 15,022 msgs/sec ~ 1.83 MB/sec (10000 msgs)
 [86] 14,985 msgs/sec ~ 1.83 MB/sec (10000 msgs)
 [87] 14,978 msgs/sec ~ 1.83 MB/sec (10000 msgs)
 [88] 14,966 msgs/sec ~ 1.83 MB/sec (10000 msgs)
 [89] 14,920 msgs/sec ~ 1.82 MB/sec (10000 msgs)
 [90] 14,874 msgs/sec ~ 1.82 MB/sec (10000 msgs)
 [91] 14,822 msgs/sec ~ 1.81 MB/sec (10000 msgs)
 [92] 14,817 msgs/sec ~ 1.81 MB/sec (10000 msgs)
 [93] 14,717 msgs/sec ~ 1.80 MB/sec (10000 msgs)
 [94] 14,715 msgs/sec ~ 1.80 MB/sec (10000 msgs)
 [95] 14,687 msgs/sec ~ 1.79 MB/sec (10000 msgs)
 [96] 14,706 msgs/sec ~ 1.80 MB/sec (10000 msgs)
 [97] 14,697 msgs/sec ~ 1.79 MB/sec (10000 msgs)
 [98] 14,692 msgs/sec ~ 1.79 MB/sec (10000 msgs)
 [99] 14,683 msgs/sec ~ 1.79 MB/sec (10000 msgs)
 [100] 14,668 msgs/sec ~ 1.79 MB/sec (10000 msgs)
 min 14,668 | avg 17,551 | max 27,027 | stddev 2,817 msgs

Old cache, cache disabled:
nats bench foo --msgs 1000000 --pub 100 --no-progress --multisubject --multisubjectmax 100000
 Pub stats: 3,794,630 msgs/sec ~ 463.21 MB/sec
 [1] 68,086 msgs/sec ~ 8.31 MB/sec (10000 msgs)
 [2] 62,718 msgs/sec ~ 7.66 MB/sec (10000 msgs)
 [3] 57,589 msgs/sec ~ 7.03 MB/sec (10000 msgs)
 [4] 56,457 msgs/sec ~ 6.89 MB/sec (10000 msgs)
 [5] 54,014 msgs/sec ~ 6.59 MB/sec (10000 msgs)
 [6] 54,073 msgs/sec ~ 6.60 MB/sec (10000 msgs)
 [7] 53,102 msgs/sec ~ 6.48 MB/sec (10000 msgs)
 [8] 52,543 msgs/sec ~ 6.41 MB/sec (10000 msgs)
 [9] 52,269 msgs/sec ~ 6.38 MB/sec (10000 msgs)
 [10] 50,392 msgs/sec ~ 6.15 MB/sec (10000 msgs)
 [11] 50,989 msgs/sec ~ 6.22 MB/sec (10000 msgs)
 [12] 50,920 msgs/sec ~ 6.22 MB/sec (10000 msgs)
 [13] 50,902 msgs/sec ~ 6.21 MB/sec (10000 msgs)
 [14] 49,691 msgs/sec ~ 6.07 MB/sec (10000 msgs)
 [15] 50,189 msgs/sec ~ 6.13 MB/sec (10000 msgs)
 [16] 49,588 msgs/sec ~ 6.05 MB/sec (10000 msgs)
 [17] 49,199 msgs/sec ~ 6.01 MB/sec (10000 msgs)
 [18] 49,178 msgs/sec ~ 6.00 MB/sec (10000 msgs)
 [19] 49,098 msgs/sec ~ 5.99 MB/sec (10000 msgs)
 [20] 48,299 msgs/sec ~ 5.90 MB/sec (10000 msgs)
 [21] 46,890 msgs/sec ~ 5.72 MB/sec (10000 msgs)
 [22] 46,690 msgs/sec ~ 5.70 MB/sec (10000 msgs)
 [23] 46,648 msgs/sec ~ 5.69 MB/sec (10000 msgs)
 [24] 45,495 msgs/sec ~ 5.55 MB/sec (10000 msgs)
 [25] 44,908 msgs/sec ~ 5.48 MB/sec (10000 msgs)
 [26] 45,436 msgs/sec ~ 5.55 MB/sec (10000 msgs)
 [27] 45,133 msgs/sec ~ 5.51 MB/sec (10000 msgs)
 [28] 45,084 msgs/sec ~ 5.50 MB/sec (10000 msgs)
 [29] 44,911 msgs/sec ~ 5.48 MB/sec (10000 msgs)
 [30] 44,080 msgs/sec ~ 5.38 MB/sec (10000 msgs)
 [31] 44,444 msgs/sec ~ 5.43 MB/sec (10000 msgs)
 [32] 43,872 msgs/sec ~ 5.36 MB/sec (10000 msgs)
 [33] 44,027 msgs/sec ~ 5.37 MB/sec (10000 msgs)
 [34] 43,986 msgs/sec ~ 5.37 MB/sec (10000 msgs)
 [35] 43,939 msgs/sec ~ 5.36 MB/sec (10000 msgs)
 [36] 43,107 msgs/sec ~ 5.26 MB/sec (10000 msgs)
 [37] 43,097 msgs/sec ~ 5.26 MB/sec (10000 msgs)
 [38] 43,652 msgs/sec ~ 5.33 MB/sec (10000 msgs)
 [39] 43,579 msgs/sec ~ 5.32 MB/sec (10000 msgs)
 [40] 43,537 msgs/sec ~ 5.31 MB/sec (10000 msgs)
 [41] 43,520 msgs/sec ~ 5.31 MB/sec (10000 msgs)
 [42] 43,498 msgs/sec ~ 5.31 MB/sec (10000 msgs)
 [43] 43,428 msgs/sec ~ 5.30 MB/sec (10000 msgs)
 [44] 43,170 msgs/sec ~ 5.27 MB/sec (10000 msgs)
 [45] 42,352 msgs/sec ~ 5.17 MB/sec (10000 msgs)
 [46] 42,222 msgs/sec ~ 5.15 MB/sec (10000 msgs)
 [47] 42,044 msgs/sec ~ 5.13 MB/sec (10000 msgs)
 [48] 41,943 msgs/sec ~ 5.12 MB/sec (10000 msgs)
 [49] 41,278 msgs/sec ~ 5.04 MB/sec (10000 msgs)
 [50] 41,719 msgs/sec ~ 5.09 MB/sec (10000 msgs)
 [51] 41,168 msgs/sec ~ 5.03 MB/sec (10000 msgs)
 [52] 41,686 msgs/sec ~ 5.09 MB/sec (10000 msgs)
 [53] 41,683 msgs/sec ~ 5.09 MB/sec (10000 msgs)
 [54] 41,622 msgs/sec ~ 5.08 MB/sec (10000 msgs)
 [55] 41,506 msgs/sec ~ 5.07 MB/sec (10000 msgs)
 [56] 41,498 msgs/sec ~ 5.07 MB/sec (10000 msgs)
 [57] 41,402 msgs/sec ~ 5.05 MB/sec (10000 msgs)
 [58] 41,333 msgs/sec ~ 5.05 MB/sec (10000 msgs)
 [59] 41,285 msgs/sec ~ 5.04 MB/sec (10000 msgs)
 [60] 40,581 msgs/sec ~ 4.95 MB/sec (10000 msgs)
 [61] 41,008 msgs/sec ~ 5.01 MB/sec (10000 msgs)
 [62] 41,004 msgs/sec ~ 5.01 MB/sec (10000 msgs)
 [63] 40,806 msgs/sec ~ 4.98 MB/sec (10000 msgs)
 [64] 40,693 msgs/sec ~ 4.97 MB/sec (10000 msgs)
 [65] 40,248 msgs/sec ~ 4.91 MB/sec (10000 msgs)
 [66] 40,738 msgs/sec ~ 4.97 MB/sec (10000 msgs)
 [67] 40,234 msgs/sec ~ 4.91 MB/sec (10000 msgs)
 [68] 40,198 msgs/sec ~ 4.91 MB/sec (10000 msgs)
 [69] 40,099 msgs/sec ~ 4.89 MB/sec (10000 msgs)
 [70] 39,722 msgs/sec ~ 4.85 MB/sec (10000 msgs)
 [71] 39,638 msgs/sec ~ 4.84 MB/sec (10000 msgs)
 [72] 39,613 msgs/sec ~ 4.84 MB/sec (10000 msgs)
 [73] 39,605 msgs/sec ~ 4.83 MB/sec (10000 msgs)
 [74] 39,598 msgs/sec ~ 4.83 MB/sec (10000 msgs)
 [75] 39,591 msgs/sec ~ 4.83 MB/sec (10000 msgs)
 [76] 39,474 msgs/sec ~ 4.82 MB/sec (10000 msgs)
 [77] 39,471 msgs/sec ~ 4.82 MB/sec (10000 msgs)
 [78] 38,903 msgs/sec ~ 4.75 MB/sec (10000 msgs)
 [79] 39,322 msgs/sec ~ 4.80 MB/sec (10000 msgs)
 [80] 39,230 msgs/sec ~ 4.79 MB/sec (10000 msgs)
 [81] 39,033 msgs/sec ~ 4.76 MB/sec (10000 msgs)
 [82] 38,992 msgs/sec ~ 4.76 MB/sec (10000 msgs)
 [83] 38,986 msgs/sec ~ 4.76 MB/sec (10000 msgs)
 [84] 38,966 msgs/sec ~ 4.76 MB/sec (10000 msgs)
 [85] 38,776 msgs/sec ~ 4.73 MB/sec (10000 msgs)
 [86] 38,741 msgs/sec ~ 4.73 MB/sec (10000 msgs)
 [87] 38,717 msgs/sec ~ 4.73 MB/sec (10000 msgs)
 [88] 38,706 msgs/sec ~ 4.72 MB/sec (10000 msgs)
 [89] 38,683 msgs/sec ~ 4.72 MB/sec (10000 msgs)
 [90] 38,562 msgs/sec ~ 4.71 MB/sec (10000 msgs)
 [91] 38,619 msgs/sec ~ 4.71 MB/sec (10000 msgs)
 [92] 38,584 msgs/sec ~ 4.71 MB/sec (10000 msgs)
 [93] 38,124 msgs/sec ~ 4.65 MB/sec (10000 msgs)
 [94] 38,574 msgs/sec ~ 4.71 MB/sec (10000 msgs)
 [95] 38,080 msgs/sec ~ 4.65 MB/sec (10000 msgs)
 [96] 38,031 msgs/sec ~ 4.64 MB/sec (10000 msgs)
 [97] 38,478 msgs/sec ~ 4.70 MB/sec (10000 msgs)
 [98] 38,471 msgs/sec ~ 4.70 MB/sec (10000 msgs)
 [99] 38,438 msgs/sec ~ 4.69 MB/sec (10000 msgs)
 [100] 38,411 msgs/sec ~ 4.69 MB/sec (10000 msgs)
 min 38,031 | avg 43,619 | max 68,086 | stddev 5,583 msgs

Old cache slCacheMax 1024:
nats bench foo --msgs 1000000 --pub 100 --no-progress --multisubject --multisubjectmax 10
 Pub stats: 13,183,781 msgs/sec ~ 1.57 GB/sec
 [1] 792,935 msgs/sec ~ 96.79 MB/sec (10000 msgs)
 [2] 416,793 msgs/sec ~ 50.88 MB/sec (10000 msgs)
 [3] 425,855 msgs/sec ~ 51.98 MB/sec (10000 msgs)
 [4] 412,288 msgs/sec ~ 50.33 MB/sec (10000 msgs)
 [5] 412,135 msgs/sec ~ 50.31 MB/sec (10000 msgs)
 [6] 425,931 msgs/sec ~ 51.99 MB/sec (10000 msgs)
 [7] 423,951 msgs/sec ~ 51.75 MB/sec (10000 msgs)
 [8] 412,812 msgs/sec ~ 50.39 MB/sec (10000 msgs)
 [9] 426,090 msgs/sec ~ 52.01 MB/sec (10000 msgs)
 [10] 412,863 msgs/sec ~ 50.40 MB/sec (10000 msgs)
 [11] 416,049 msgs/sec ~ 50.79 MB/sec (10000 msgs)
 [12] 412,533 msgs/sec ~ 50.36 MB/sec (10000 msgs)
 [13] 411,879 msgs/sec ~ 50.28 MB/sec (10000 msgs)
 [14] 329,393 msgs/sec ~ 40.21 MB/sec (10000 msgs)
 [15] 330,119 msgs/sec ~ 40.30 MB/sec (10000 msgs)
 [16] 292,080 msgs/sec ~ 35.65 MB/sec (10000 msgs)
 [17] 291,857 msgs/sec ~ 35.63 MB/sec (10000 msgs)
 [18] 285,308 msgs/sec ~ 34.83 MB/sec (10000 msgs)
 [19] 291,834 msgs/sec ~ 35.62 MB/sec (10000 msgs)
 [20] 285,844 msgs/sec ~ 34.89 MB/sec (10000 msgs)
 [21] 278,464 msgs/sec ~ 33.99 MB/sec (10000 msgs)
 [22] 278,536 msgs/sec ~ 34.00 MB/sec (10000 msgs)
 [23] 284,219 msgs/sec ~ 34.69 MB/sec (10000 msgs)
 [24] 277,294 msgs/sec ~ 33.85 MB/sec (10000 msgs)
 [25] 275,357 msgs/sec ~ 33.61 MB/sec (10000 msgs)
 [26] 275,334 msgs/sec ~ 33.61 MB/sec (10000 msgs)
 [27] 275,245 msgs/sec ~ 33.60 MB/sec (10000 msgs)
 [28] 268,392 msgs/sec ~ 32.76 MB/sec (10000 msgs)
 [29] 264,920 msgs/sec ~ 32.34 MB/sec (10000 msgs)
 [30] 259,936 msgs/sec ~ 31.73 MB/sec (10000 msgs)
 [31] 252,694 msgs/sec ~ 30.85 MB/sec (10000 msgs)
 [32] 245,938 msgs/sec ~ 30.02 MB/sec (10000 msgs)
 [33] 231,012 msgs/sec ~ 28.20 MB/sec (10000 msgs)
 [34] 246,502 msgs/sec ~ 30.09 MB/sec (10000 msgs)
 [35] 242,007 msgs/sec ~ 29.54 MB/sec (10000 msgs)
 [36] 224,047 msgs/sec ~ 27.35 MB/sec (10000 msgs)
 [37] 793,469 msgs/sec ~ 96.86 MB/sec (10000 msgs)
 [38] 767,256 msgs/sec ~ 93.66 MB/sec (10000 msgs)
 [39] 766,996 msgs/sec ~ 93.63 MB/sec (10000 msgs)
 [40] 232,576 msgs/sec ~ 28.39 MB/sec (10000 msgs)
 [41] 237,156 msgs/sec ~ 28.95 MB/sec (10000 msgs)
 [42] 214,475 msgs/sec ~ 26.18 MB/sec (10000 msgs)
 [43] 210,776 msgs/sec ~ 25.73 MB/sec (10000 msgs)
 [44] 187,124 msgs/sec ~ 22.84 MB/sec (10000 msgs)
 [45] 193,815 msgs/sec ~ 23.66 MB/sec (10000 msgs)
 [46] 193,851 msgs/sec ~ 23.66 MB/sec (10000 msgs)
 [47] 193,652 msgs/sec ~ 23.64 MB/sec (10000 msgs)
 [48] 196,220 msgs/sec ~ 23.95 MB/sec (10000 msgs)
 [49] 193,808 msgs/sec ~ 23.66 MB/sec (10000 msgs)
 [50] 186,851 msgs/sec ~ 22.81 MB/sec (10000 msgs)
 [51] 193,518 msgs/sec ~ 23.62 MB/sec (10000 msgs)
 [52] 196,275 msgs/sec ~ 23.96 MB/sec (10000 msgs)
 [53] 196,566 msgs/sec ~ 23.99 MB/sec (10000 msgs)
 [54] 193,770 msgs/sec ~ 23.65 MB/sec (10000 msgs)
 [55] 196,147 msgs/sec ~ 23.94 MB/sec (10000 msgs)
 [56] 193,764 msgs/sec ~ 23.65 MB/sec (10000 msgs)
 [57] 185,355 msgs/sec ~ 22.63 MB/sec (10000 msgs)
 [58] 183,785 msgs/sec ~ 22.43 MB/sec (10000 msgs)
 [59] 189,493 msgs/sec ~ 23.13 MB/sec (10000 msgs)
 [60] 186,976 msgs/sec ~ 22.82 MB/sec (10000 msgs)
 [61] 177,384 msgs/sec ~ 21.65 MB/sec (10000 msgs)
 [62] 184,658 msgs/sec ~ 22.54 MB/sec (10000 msgs)
 [63] 166,887 msgs/sec ~ 20.37 MB/sec (10000 msgs)
 [64] 174,228 msgs/sec ~ 21.27 MB/sec (10000 msgs)
 [65] 164,030 msgs/sec ~ 20.02 MB/sec (10000 msgs)
 [66] 169,250 msgs/sec ~ 20.66 MB/sec (10000 msgs)
 [67] 168,647 msgs/sec ~ 20.59 MB/sec (10000 msgs)
 [68] 166,539 msgs/sec ~ 20.33 MB/sec (10000 msgs)
 [69] 167,785 msgs/sec ~ 20.48 MB/sec (10000 msgs)
 [70] 165,580 msgs/sec ~ 20.21 MB/sec (10000 msgs)
 [71] 164,850 msgs/sec ~ 20.12 MB/sec (10000 msgs)
 [72] 163,736 msgs/sec ~ 19.99 MB/sec (10000 msgs)
 [73] 162,737 msgs/sec ~ 19.87 MB/sec (10000 msgs)
 [74] 156,067 msgs/sec ~ 19.05 MB/sec (10000 msgs)
 [75] 161,883 msgs/sec ~ 19.76 MB/sec (10000 msgs)
 [76] 159,926 msgs/sec ~ 19.52 MB/sec (10000 msgs)
 [77] 160,657 msgs/sec ~ 19.61 MB/sec (10000 msgs)
 [78] 160,201 msgs/sec ~ 19.56 MB/sec (10000 msgs)
 [79] 160,050 msgs/sec ~ 19.54 MB/sec (10000 msgs)
 [80] 159,953 msgs/sec ~ 19.53 MB/sec (10000 msgs)
 [81] 151,032 msgs/sec ~ 18.44 MB/sec (10000 msgs)
 [82] 146,681 msgs/sec ~ 17.91 MB/sec (10000 msgs)
 [83] 145,443 msgs/sec ~ 17.75 MB/sec (10000 msgs)
 [84] 139,811 msgs/sec ~ 17.07 MB/sec (10000 msgs)
 [85] 144,737 msgs/sec ~ 17.67 MB/sec (10000 msgs)
 [86] 137,836 msgs/sec ~ 16.83 MB/sec (10000 msgs)
 [87] 142,750 msgs/sec ~ 17.43 MB/sec (10000 msgs)
 [88] 142,889 msgs/sec ~ 17.44 MB/sec (10000 msgs)
 [89] 140,320 msgs/sec ~ 17.13 MB/sec (10000 msgs)
 [90] 136,593 msgs/sec ~ 16.67 MB/sec (10000 msgs)
 [91] 138,869 msgs/sec ~ 16.95 MB/sec (10000 msgs)
 [92] 140,343 msgs/sec ~ 17.13 MB/sec (10000 msgs)
 [93] 134,802 msgs/sec ~ 16.46 MB/sec (10000 msgs)
 [94] 134,397 msgs/sec ~ 16.41 MB/sec (10000 msgs)
 [95] 134,347 msgs/sec ~ 16.40 MB/sec (10000 msgs)
 [96] 137,316 msgs/sec ~ 16.76 MB/sec (10000 msgs)
 [97] 137,230 msgs/sec ~ 16.75 MB/sec (10000 msgs)
 [98] 136,996 msgs/sec ~ 16.72 MB/sec (10000 msgs)
 [99] 132,765 msgs/sec ~ 16.21 MB/sec (10000 msgs)
 [100] 132,286 msgs/sec ~ 16.15 MB/sec (10000 msgs)
 min 132,286 | avg 246,746 | max 793,469 | stddev 139,200 msgs
</details>

<details>
 <summary>New Cache</summary>

New cache slCacheMax 1024:
nats bench foo --msgs 1000000 --pub 100 --no-progress --multisubject --multisubjectmax 100000

Pub stats: 4,390,717 msgs/sec ~ 535.98 MB/sec
 [1] 53,083 msgs/sec ~ 6.48 MB/sec (10000 msgs)
 [2] 52,676 msgs/sec ~ 6.43 MB/sec (10000 msgs)
 [3] 50,037 msgs/sec ~ 6.11 MB/sec (10000 msgs)
 [4] 49,132 msgs/sec ~ 6.00 MB/sec (10000 msgs)
 [5] 49,100 msgs/sec ~ 5.99 MB/sec (10000 msgs)
 [6] 48,791 msgs/sec ~ 5.96 MB/sec (10000 msgs)
 [7] 48,110 msgs/sec ~ 5.87 MB/sec (10000 msgs)
 [8] 48,403 msgs/sec ~ 5.91 MB/sec (10000 msgs)
 [9] 48,333 msgs/sec ~ 5.90 MB/sec (10000 msgs)
 [10] 48,214 msgs/sec ~ 5.89 MB/sec (10000 msgs)
 [11] 47,297 msgs/sec ~ 5.77 MB/sec (10000 msgs)
 [12] 47,717 msgs/sec ~ 5.82 MB/sec (10000 msgs)
 [13] 47,713 msgs/sec ~ 5.82 MB/sec (10000 msgs)
 [14] 47,503 msgs/sec ~ 5.80 MB/sec (10000 msgs)
 [15] 47,423 msgs/sec ~ 5.79 MB/sec (10000 msgs)
 [16] 47,391 msgs/sec ~ 5.79 MB/sec (10000 msgs)
 [17] 47,277 msgs/sec ~ 5.77 MB/sec (10000 msgs)
 [18] 47,270 msgs/sec ~ 5.77 MB/sec (10000 msgs)
 [19] 47,056 msgs/sec ~ 5.74 MB/sec (10000 msgs)
 [20] 46,520 msgs/sec ~ 5.68 MB/sec (10000 msgs)
 [21] 47,018 msgs/sec ~ 5.74 MB/sec (10000 msgs)
 [22] 46,437 msgs/sec ~ 5.67 MB/sec (10000 msgs)
 [23] 46,187 msgs/sec ~ 5.64 MB/sec (10000 msgs)
 [24] 46,663 msgs/sec ~ 5.70 MB/sec (10000 msgs)
 [25] 46,136 msgs/sec ~ 5.63 MB/sec (10000 msgs)
 [26] 46,642 msgs/sec ~ 5.69 MB/sec (10000 msgs)
 [27] 46,567 msgs/sec ~ 5.68 MB/sec (10000 msgs)
 [28] 46,530 msgs/sec ~ 5.68 MB/sec (10000 msgs)
 [29] 46,499 msgs/sec ~ 5.68 MB/sec (10000 msgs)
 [30] 45,926 msgs/sec ~ 5.61 MB/sec (10000 msgs)
 [31] 46,474 msgs/sec ~ 5.67 MB/sec (10000 msgs)
 [32] 46,465 msgs/sec ~ 5.67 MB/sec (10000 msgs)
 [33] 46,445 msgs/sec ~ 5.67 MB/sec (10000 msgs)
 [34] 46,439 msgs/sec ~ 5.67 MB/sec (10000 msgs)
 [35] 45,902 msgs/sec ~ 5.60 MB/sec (10000 msgs)
 [36] 46,444 msgs/sec ~ 5.67 MB/sec (10000 msgs)
 [37] 46,374 msgs/sec ~ 5.66 MB/sec (10000 msgs)
 [38] 46,351 msgs/sec ~ 5.66 MB/sec (10000 msgs)
 [39] 46,269 msgs/sec ~ 5.65 MB/sec (10000 msgs)
…